+++ src/utility.py
@@ -4,6 +4,7 @@ import os
 import sys
 import logging
 
+
 def get_logger(name):
     logging.basicConfig(
         format='%(asctime)s - %(name)s - %(levelname)s %(threadName)s:%(funcName)s: %(message)s',
@@ -12,6 +13,7 @@ def get_logger(name):
     logger = logging.getLogger(name)
     logger.setLevel(logging.INFO)
     return logger
+
 
 @dataclass
 class TimeInterval:
@@ -31,6 +33,13 @@ class TimeInterval:
     def has_end(self):
         return self.end != inf
 
+
 def restart_program():
     python = sys.executable
     os.execl(python, python, * sys.argv)
+
+
+def make_even(x):
+    if x % 2 == 1:
+        x -= 1
+    return x
